debug

+++ TEST/TagReadWithGeneExon/test/scripts/test.py
@@ -40,7 +40,7 @@ for ref in reads_dict:
     for read in reads_list:
         total += len(read.get_blocks())
 
-    R = np.zeros(total, dtype='int16')
+    R = np.zeros(total)
     B = np.arange(total)
 
     ctr = 0
